cnn_cfar/requirements.txt
fastapi
uvicorn 
python-multipart
pillow
tensorflow>=2.19.0